+++ NEWS.md
@@ -12,6 +12,13 @@ as R code for the official GitHub language statistics calculated via
 
 [linguist]: https://github.com/github/linguist
 
+* Address [callr 3.3.0 bug][callr-bug-3.3.0] that writes objects to the global
+environment, causing the workflowr check of the global environment to fail. The
+failed check now explains that the problem can be fixed by updating the callr
+package.
+
+[callr-bug-3.3.0]: https://github.com/r-lib/callr/commit/9f7665e1081da6f5134b214da694b4461d05659f
+
 # workflowr 1.4.0
 
 This minor release of workflowr features further GitHub integration, a new